chore: move meta handling for sync
